Make "moved indices" track item movement after deletions and insertions
This change is to make BKDelta have indices that are compatible with UITableView. Specifically, UITableView's batching algorithm processes:

  1. reloads (using the old indices)
  2. deletions (old indices)
  3. insertions (new indices)

Note that moving cells is left out. So if you wanted to apply a full delta without calling `reloadData` and blowing away everything, you could do:

  -- start batched update --
  reloads
  deletions
  insertions
  -- end batched update --
  moves

This doesn't totally work yet since the moves don't account for prior moves, but it avoids filling the `movedIndices` array if deletions and insertions put the old array in the new array's state.